pyproject.toml
[project]
name = "django-tortoise-objects"
version = "0.1.0"
description = "Bridge Django ORM models to Tortoise ORM for truly async database access"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "Andrei Bondar"},
]
keywords = ["django", "tortoise-orm", "async", "orm", "database"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "django>=4.2",
    "tortoise-orm>=1.1.2",
]

[project.urls]
Homepage = "https://github.com/tortoise/django-tortoise-objects"
Repository = "https://github.com/tortoise/django-tortoise-objects"
Issues = "https://github.com/tortoise/django-tortoise-objects/issues"

[project.optional-dependencies]
pg = ["psycopg[binary]>=3.1"]
mysql = ["asyncmy>=0.2"]
sqlite = ["aiosqlite>=0.19"]
all = ["psycopg[binary]>=3.1", "asyncmy>=0.2", "aiosqlite>=0.19"]
dev = [
    "pytest>=7.0",
    "pytest-django>=4.5",
    "pytest-asyncio>=0.21",
    "aiosqlite>=0.19",
    "ruff>=0.9",
    "mypy>=1.14",
    "django-stubs>=5.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["django_tortoise"]

# -- pytest ------------------------------------------------------------------

[tool.pytest.ini_options]
asyncio_mode = "auto"

# -- ruff --------------------------------------------------------------------

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "SIM",   # flake8-simplify
    "RUF",   # ruff-specific rules
]
ignore = [
    "E501",    # line too long (formatter handles this)
    "UP007",   # Use X | Y for union types (already used, but keep for 3.10 compat)
    "RUF012",  # mutable ClassVar — false positives on Django Meta classes
    "SIM108",  # ternary operator — if/else is often clearer with comments
]

[tool.ruff.lint.isort]
known-first-party = ["django_tortoise"]

# -- mypy --------------------------------------------------------------------

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
plugins = ["mypy_django_plugin.main"]

[[tool.mypy.overrides]]
module = [
    "tortoise.*",
]
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "tests.django_settings"
